Add files via upload
Interface Design Using CSSMain ProjectSelling Out!Due: Week 13Description:Using everything we have covered over the course of the semester you willneed to build a small website with the purpose of selling a product. The productcan be of your choosing.Instructions:Please read and follow the instructions below carefully:Design:1. Decide on the product that you wish to sell. This can be any product realor fake!2. Gather any images or text that will be needed for the design & website.3. Create an accurate wireframe/design for the homepage (the index file).Build (HTML):1. The page should be created using the appropriate semantic elementsand should be error free when validated.2. The following pages will be required:a. Homepageb. About pagec. Product paged. Contact page3. All pages should be responsive on all the major viewports.4. The pages should have a:a. Header that includes a logo and navigationb. A call to action (on at least the homepage)c. Footer that contains any company information and/or contactform.Style It (CSS):
1. The website should have a professional finish.2. The CSS should be error free when validated.3. The color scheme should complement the product and follow one of thefollowing:a. Monochromeb. Complementaryc. Analogousd. Or Triadic4. The text should have at least 2 types of font families.5. Letter spacing or kerning should be used where appropriate.6. The pages should have a healthy number of different effects andenhancements, some ideas:a. Hover effectsb. Background video / imagesc. SVG’sd. Filterse. Etc.7. The layout should be engaging and easy to follow